When creating a new branch off of a main branch that is behind its upstream, use the upstream instead
Very often my local main branch is behind its upstream, because I have no reason
to keep it up to date. In the olden days it used to be necessary to keep it up
to date in order to rebase feature branches onto it, so you'd have to press `f`
on it occasionally before pressing `r`. This is no longer necessary now that we
have the "rebase onto base branch" command (`r b`).

The only time I now need an up-to-date local main is when I want to create a new
branch off of it. This is annoying, because I have to press `f` and wait for it
to complete before I can create a new branch; but I know that origin/main is
probably up to date (or close to up to date), because lazygit fetches it in the
background all the time. So why can't I just create my new branch off of
origin/main, then? (Sure, I could switch over to the Remotes tab, open origin,
and select origin/main there, but that's cumbersome.)

So this is exactly what this commit does: if the local main is strictly behind
its upstream (as opposed to diverged, which should be very rare), then when
pressing `n` on it we create the new branch off of origin/main.