Switch order of rotate and return for next allBranchesLogCmd
This made some amount of sense in the previous world. If users were
using `statusPanelView: dashboard`, then it would make sense for them to
go to their first log when they switch over from the dashboard to the
log commands by pressing `a`. This doesn't end up working out super well
though for users of `statusPanelView: allBranchesLog`, where upon
reaching the view with `1`, then would press `a` and see nothing.
Upon returning to the view with a `2` and then `1`, they would then see
the new value!

Overall, it seems better that the command that you are viewing the
output of is still the index of the command stored in lazygit's state.
If we want to add more code to make it so that lazygit knows it is _not
yet_ showing a branch command, and is instead on the dashboard, and make
it so that it doesn't rotate at all in that case, we could do that. But
that seems like too much work that no user would ever appreciate.

I am half assuming here that users who love allBranchesLogCmds, and use
multiple of them, are also running with `statusPanelView:
allBranchesLog`, so they will not be negatively impacted by this.