Set explicit mainSectionWeight when forcing main view to full width
Previously we were relying on setting just the sideSectionWeight in order
to ensure the main panel was fully focused. This can cause a problem if
the user had their `gui.sidePanelWidth` set to 1. Such a value make the
mainPanelWidth = 0 by default. By just setting the sideSectionWeight, you
then communicate to the UI that you want _both_ halves of the screen to have
0 width! No good!

This change seems like it won't impact users of more standard
`gui.sidePanelWidth` values, so it's just a free win!