Make stash loading work in filtering mode
This broke with the introduction of the age in the stashes list in bc330b8ff3ef
(which was included in 0.41, so 12 minor versions ago).

This makes it work again when filtering by file, but it still doesn't work when
filtering by folder (and never has). We'll fix that next.